// test/page-objects/types.ts
/**
 * Represents a contact with personal information.
 */
export interface Contact {
    firstName: string;
    lastName: string;
    birthdate: string;
    email: string;
    phone: string;
    street1: string;
    street2: string;
    city: string;
    stateProvince: string;
    postalCode: string;
    country: string;
}